+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   `externalSymbolLinkMappings` now uses the TypeScript reported link target if available, #2725.
 -   TypeDoc will no longer omit the modules page if a project contains only modules/documents, #2730.
+-   Fixed missing breadcrumbs on project page, #2728.
 
 ## v0.26.8 (2024-10-04)