+++ CHANGELOG.md
@@ -1,10 +1,21 @@
 # Unreleased
 
+### Features
+
+-   Added `headings` option to control optional headings, #2729.
+-   Added a folder icon to page navigation elements which are not links, #2741.
+
 ### Bug Fixes
 
 -   `externalSymbolLinkMappings` now uses the TypeScript reported link target if available, #2725.
 -   TypeDoc will no longer omit the modules page if a project contains only modules/documents, #2730.
 -   Fixed missing breadcrumbs on project page, #2728.
+-   TypeDoc will no longer render an empty readme page if no readme was found.
+
+### Thanks!
+
+-   @lriggle-strib
+-   @mrfigg
 
 ## v0.26.8 (2024-10-04)